+++ src/ble_module.c
@@ -1,0 +1,187 @@
+#include <zephyr/kernel.h>
+#include <zephyr/logging/log.h>
+
+#include <zephyr/bluetooth/bluetooth.h>
+#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/conn.h>
+#include <zephyr/bluetooth/uuid.h>
+#include <zephyr/bluetooth/gatt.h>
+#include <zephyr/bluetooth/services/bas.h>
+#include <zephyr/bluetooth/services/hrs.h>
+#include <zephyr/sys/ring_buffer.h>
+
+#include <zephyr/settings/settings.h>
+
+#include "hpi_data_service.h"
+#include "cmd_module.h"
+
+#define LOG_LEVEL CONFIG_LOG_DEFAULT_LEVEL
+LOG_MODULE_REGISTER(ble_module);
+
+struct bt_conn *current_conn;
+
+static const struct bt_data ad[] = {
+
+	BT_DATA_BYTES(BT_DATA_FLAGS, (BT_LE_AD_GENERAL | BT_LE_AD_NO_BREDR)),
+	BT_DATA_BYTES(BT_DATA_UUID16_ALL,
+				  BT_UUID_16_ENCODE(BT_UUID_HRS_VAL),
+				  BT_UUID_16_ENCODE(BT_UUID_BAS_VAL),
+				  BT_UUID_16_ENCODE(BT_UUID_DIS_VAL))};
+
+static void connected(struct bt_conn *conn, uint8_t err)
+{
+	if (err)
+	{
+		printk("Connection failed (err 0x%02x)\n", err);
+	}
+	else
+	{
+		printk("Connected\n");
+		send_status_serial(BLE_STATUS_CONNECTED);
+		current_conn = bt_conn_ref(conn);
+	}
+}
+
+static void disconnected(struct bt_conn *conn, uint8_t reason)
+{
+	printk("Disconnected (reason 0x%02x)\n", reason);
+	send_status_serial(BLE_STATUS_DISCONNECTED);
+	if (current_conn)
+	{
+		bt_conn_unref(current_conn);
+		current_conn = NULL;
+	}
+}
+
+static void security_changed(struct bt_conn *conn, bt_security_t level,
+							 enum bt_security_err err)
+{
+	char addr[BT_ADDR_LE_STR_LEN];
+
+	bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
+
+	if (!err)
+	{
+		printk("Security changed: %s level %u\n", addr, level);
+	}
+	else
+	{
+		printk("Security failed: %s level %u err %d\n", addr, level,
+			   err);
+	}
+}
+
+BT_CONN_CB_DEFINE(conn_callbacks) = {
+	.connected = connected,
+	.disconnected = disconnected,
+	.security_changed = security_changed,
+};
+
+static void bt_ready(void)
+{
+	int err;
+
+	err = hpi_service_init();
+
+	printk("Bluetooth initialized\n");
+
+	if (IS_ENABLED(CONFIG_SETTINGS))
+	{
+		settings_load();
+	}
+
+	err = bt_le_adv_start(BT_LE_ADV_CONN_NAME, ad, ARRAY_SIZE(ad), NULL, 0);
+	if (err)
+	{
+		printk("Advertising failed to start (err %d)\n", err);
+		return;
+	}
+
+	printk("Advertising successfully started\n");
+}
+
+static void auth_passkey_confirm(struct bt_conn *conn, unsigned int passkey)
+{
+	char passkey_str[7];
+	char addr[BT_ADDR_LE_STR_LEN];
+
+	bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
+
+	snprintk(passkey_str, ARRAY_SIZE(passkey_str), "%06u", passkey);
+
+	printk("Passkey for %s: %s\n", addr, passkey_str);
+
+	// k_poll_signal_raise(&passkey_enter_signal, 0);
+}
+
+static void auth_passkey_display(struct bt_conn *conn, unsigned int passkey)
+{
+	char passkey_str[7];
+	char addr[BT_ADDR_LE_STR_LEN];
+
+	bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
+
+	snprintk(passkey_str, ARRAY_SIZE(passkey_str), "%06u", passkey);
+
+	LOG_DBG("Passkey for %s: %s", addr, passkey_str);
+}
+
+static void auth_cancel(struct bt_conn *conn)
+{
+	char addr[BT_ADDR_LE_STR_LEN];
+
+	bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
+
+	printk("Pairing cancelled: %s\n", addr);
+}
+
+static struct bt_conn_auth_cb auth_cb_display = {
+	.cancel = auth_cancel,
+	.passkey_display = auth_passkey_display,
+	.passkey_confirm = auth_passkey_confirm,
+};
+
+void remove_separators(char *str)
+{
+	char *pr = str, *pw = str;
+	char c = ':';
+	while (*pr)
+	{
+		*pw = *pr++;
+		pw += (*pw != c);
+	}
+	*pw = '\0';
+}
+
+void set_ble_name()
+{
+	bt_addr_le_t addrs;
+	size_t count;
+
+	char addr_str[50];
+	size_t add_str_len;
+	char dev_name[30] = "WISER-";
+
+	bt_id_get(&addrs, &count);
+
+	bt_addr_le_to_str(&addrs, addr_str, 10);
+	remove_separators(addr_str);
+	strcat(dev_name, addr_str);
+	bt_set_name(dev_name);
+}
+
+void ble_module_init(){
+    int err=0;
+    
+    err = bt_enable(NULL);
+	if (err)
+	{
+		LOG_ERR("Bluetooth init failed (err %d)", err);
+		return;
+	}
+
+	bt_ready();
+	set_ble_name();
+
+	bt_conn_auth_cb_register(&auth_cb_display);
+}
