+++ src/ble_module.h
@@ -1,0 +1,1 @@
+void ble_module_init();
